[Allure-ee] Change alb routing rules, update requirementsdependency repository
Summary: Remove secret resource because we want to be able to specify a custom secret resource (allow SealedSecret).

In addition to this,we are changing the alb ingress rule to allow /*, as otherwise it will 403 when we attempt to access uaa endpoints and such.

Test Plan: Sync to argocd and confirm changes